+++ ui_components.py
@@ -1166,14 +1166,10 @@ class HelpView(discord.ui.View):
             embed.add_field(name='Parameters', value=param_str, inline=False)
             embed.add_field(name='Usage', value='This is a slash command. Type / to get the command', inline=False)
         elif isinstance(cmd, discord.SlashCommandGroup):
-            embed.add_field(name='Subcommands',
-                            value=f'`{" ".join(f"`{subcmd.name}`" for subcmd in cmd.walk_commands())}`', inline=False)
-            embed.add_field(name='Usage',
-                            value=f'This is a Slash Command Group. Type /{cmd.name} to get the subcommands')
+            embed.add_field(name='Subcommands', value=f'{" ".join(f"`{subcmd.name}`" for subcmd in cmd.walk_commands())}', inline=False)
+            embed.add_field(name='Usage', value=f'This is a Slash Command Group. Type /{cmd.name} to get the subcommands')
         else:
-            embed.add_field(name='Usage',
-                            value='This is an application command. Right click on a message/user to get the command',
-                            inline=False)
+            embed.add_field(name='Usage', value='This is an application command. Right click on a message/user to get the command', inline=False)
 
         await interaction.response.edit_message(content=None, embed=embed, view=None)
         self.stop()